+++ examples/rust/src/bin/client.rs
@@ -6,7 +6,9 @@ use {
     indicatif::{MultiProgress, ProgressBar, ProgressStyle},
     log::{error, info},
     serde_json::{json, Value},
-    solana_sdk::{hash::Hash, pubkey::Pubkey, signature::Signature},
+    solana_hash::Hash,
+    solana_pubkey::Pubkey,
+    solana_signature::Signature,
     solana_transaction_status::UiTransactionEncoding,
     std::{
         collections::HashMap,
@@ -21,6 +23,7 @@ use {
     yellowstone_grpc_client::{GeyserGrpcClient, GeyserGrpcClientError, Interceptor},
     yellowstone_grpc_proto::{
         convert_from,
+        geyser::SlotStatus,
         plugin::filter::message::FilteredUpdate,
         prelude::{
             subscribe_request_filter_accounts_filter::Filter as AccountsFilterOneof,
@@ -193,6 +196,7 @@ enum Action {
     HealthCheck,
     HealthWatch,
     Subscribe(Box<ActionSubscribe>),
+    SubscribeReplayInfo,
     Ping {
         #[clap(long, short, default_value_t = 0)]
         count: i32,
@@ -255,6 +259,10 @@ struct ActionSubscribe {
     /// Filter slots by commitment
     #[clap(long)]
     slots_filter_by_commitment: bool,
+
+    /// Subscribe on interslot slot updates
+    #[clap(long)]
+    slots_interslot_updates: bool,
 
     /// Subscribe on transactions updates
     #[clap(long)]
@@ -451,6 +459,7 @@ impl Action {
                         "client".to_owned(),
                         SubscribeRequestFilterSlots {
                             filter_by_commitment: Some(args.slots_filter_by_commitment),
+                            interslot_updates: Some(args.slots_interslot_updates),
                         },
                     );
                 }
@@ -598,6 +607,11 @@ async fn main() -> anyhow::Result<()> {
 
                     geyser_subscribe(client, request, resub, stats, verify_encoding).await
                 }
+                Action::SubscribeReplayInfo => client
+                    .subscribe_replay_info()
+                    .await
+                    .map_err(anyhow::Error::new)
+                    .map(|response| info!("response: {response:?}")),
                 Action::Ping { count } => client
                     .ping(*count)
                     .await
@@ -772,7 +786,7 @@ async fn geyser_subscribe(
                         print_update("account", created_at, &filters, value);
                     }
                     Some(UpdateOneof::Slot(msg)) => {
-                        let status = CommitmentLevel::try_from(msg.status)
+                        let status = SlotStatus::try_from(msg.status)
                             .context("failed to decode commitment")?;
                         print_update(
                             "slot",