+++ examples/rust/src/bin/subscribe-ping.rs
@@ -44,12 +44,17 @@ async fn main() -> anyhow::Result<()> {
     futures::try_join!(
         async move {
             subscribe_tx
-            .send(SubscribeRequest {
-                slots: maplit::hashmap! { "".to_owned() => SubscribeRequestFilterSlots { filter_by_commitment: Some(true) } },
-                commitment: Some(CommitmentLevel::Processed as i32),
-                ..Default::default()
-            })
-            .await?;
+                .send(SubscribeRequest {
+                    slots: maplit::hashmap! {
+                        "".to_owned() => SubscribeRequestFilterSlots {
+                            filter_by_commitment: Some(true),
+                            interslot_updates: Some(false)
+                        }
+                    },
+                    commitment: Some(CommitmentLevel::Processed as i32),
+                    ..Default::default()
+                })
+                .await?;
 
             let mut timer = interval(Duration::from_secs(3));
             let mut id = 0;